restaurant_project/requirements.txt
Django>=6.0,<7.0
django-allauth>=65.0.0
django-cleanup>=9.0.0
Pillow>=11.0.0
